+++ docs/changelog.md
@@ -1,5 +1,11 @@
 # rcc change log
 
+## v11.3.4 (date: 11.10.2021)
+
+- new toplevel flag to turn on `--strict` environment handling, and for now
+  this make rcc to run `pip check` after environment install completes
+- added timeout to metrics sending
+
 ## v11.3.3 (date: 8.10.2021)
 
 - micromamba update to version 0.16.0